requirements.txt
garminconnect>=0.2.28
gspread>=6.0.0
google-auth>=2.29.0
